sharepoint-salsify-integration/requirements.txt
msal>=1.30.0,<2.0.0
requests>=2.32.0,<3.0.0
pydantic>=2.8.0,<3.0.0
PyYAML>=6.0.1,<7.0.0
python-dotenv>=1.0.1,<2.0.0
prometheus-client>=0.20.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
fastapi>=0.111.0,<1.0.0
tenacity>=9.0.0,<10.0.0
pytest>=8.2.0,<9.0.0
